requirements.txt
playwright
google-genai
python-dotenv
colorama
beautifulsoup4
streamlit
watchdog
